Hotel Parser
Began with reading in yaml file and creating a hotel object which makes use of a Struct object. Struct was a convenient way to bundle the attributes in the yaml file into an array, and allows for method calls and the use of the [ ] operator in addition to the use of .first, and .last (as shown in the two README examples). Two helper methods were created in order to traverse both the hotels, and their rooms respectively. I faced some issues with File.expand_path so I had to slightly tweak the way in which the argument was passed to 'hotel_parser_spec.rb'. If possible, I would love to discuss why this might be happening as I tried to trouble shoot this for some time.